requirements.txt
amqp==2.1.4
anyjson==0.3.3
appdirs==1.4.0
billiard==3.5.0.2
celery==4.0.2
Django==1.10.4
django-allauth==0.29.0
django-celery==3.2.1
django-celery-beat==1.0.1
django-mssql==1.8
django-ratelimit==1.0.1
django-rest-auth==0.8.2
djangorestframework==3.5.3
gunicorn==19.6.0
kombu==4.0.2
MySQL-python==1.2.3
oauthlib==2.0.1
packaging==16.8
psycopg2==2.6.2
pyparsing==2.1.10
python-openid==2.2.5
pytz==2016.10
requests==2.12.4
requests-oauthlib==0.7.0
six==1.10.0
vine==1.1.3
